+++ src/components/User/UserAvatar.tsx
@@ -4,14 +4,19 @@ import clsx from 'clsx';
 import { UserReport } from '../../types/common';
 import styles from './UserAvatar.module.css';
 import { useDataDragonVersion } from '../../hooks/useReport';
+import IconButton from '../common/IconButton';
+import StarIcon from '../icons/StarIcon';
+import { usePrimaryUserID } from '../../hooks/usePreferences';
 
 interface Props {
+    id: string;
     variant: 'heading';
     user: UserReport;
     index: number;
 }
-export default function UserAvatar({ variant, user, index }: Props) {
+export default function UserAvatar({ variant, user, index, id }: Props) {
     const dataDragonVersion = useDataDragonVersion();
+    const [primaryUserID, setPrimaryUserID] = usePrimaryUserID();
     const [summonerIndex, setSummonerIndex] = useState(0);
     useEffect(() => {
         const interval = setInterval(() => {
@@ -19,6 +24,9 @@ export default function UserAvatar({ variant, user, index }: Props) {
         }, 60_000);
         return () => clearInterval(interval);
     });
+    const handleSetPrimary = () => {
+        setPrimaryUserID(id);
+    };
     return (
         <div className={styles.container}>
             {user.Summoners.map((summoner, i) => (
@@ -40,6 +48,15 @@ export default function UserAvatar({ variant, user, index }: Props) {
                     >
                         {summoner.Level}
                     </code>
+                    <div
+                        className={clsx(styles.primaryToggle, {
+                            [styles.isPrimary]: primaryUserID === id,
+                        })}
+                    >
+                        <IconButton active={primaryUserID === id} onClick={handleSetPrimary}>
+                            <StarIcon />
+                        </IconButton>
+                    </div>
                     <div className={styles.name}>{summoner.Name}</div>
                 </div>
             ))}