+++ src/utils.ts
@@ -1,3 +1,5 @@
+import { AtomEffect } from 'recoil';
+
 export const numberFormat = new Intl.NumberFormat(undefined, {
     maximumSignificantDigits: 4,
     minimumSignificantDigits: 4,
@@ -10,9 +12,7 @@ export function formatNumbers(...numbers: number[]): string[] {
     for (let i = 1; i < 9; i++) {
         if (min < 1000 ** i) {
             return numbers.map(
-                (n) =>
-                    numberFormat.format(n / 1000 ** (i - 1)) +
-                    NUMBER_SUFFIXES[i - 1]
+                (n) => numberFormat.format(n / 1000 ** (i - 1)) + NUMBER_SUFFIXES[i - 1]
             );
         }
     }
@@ -30,10 +30,24 @@ export function formatDurations(...durations: number[]): string[] {
     const min = Math.min(...durations);
     for (const [suffix, seconds] of Object.entries(DURATION_SUFFIXES)) {
         if (min >= seconds) {
-            return durations.map(
-                (d) => numberFormat.format(d / seconds) + ' ' + suffix
-            );
+            return durations.map((d) => numberFormat.format(d / seconds) + ' ' + suffix);
         }
     }
-    return durations.map(d => d < 1 ? 'No time' : d + ' seconds');
+    return durations.map((d) => (d < 1 ? 'No time' : d + ' seconds'));
+}
+
+export function localStorageEffect<T>(key: string): AtomEffect<T> {
+    const storageKey = `lol-stats-${key}`;
+    return ({ setSelf, onSet }) => {
+        const savedValue = localStorage.getItem(storageKey);
+        if (savedValue != null) {
+            setSelf(JSON.parse(savedValue));
+        }
+
+        onSet((newValue, _, isReset) => {
+            isReset
+                ? localStorage.removeItem(storageKey)
+                : localStorage.setItem(storageKey, JSON.stringify(newValue));
+        });
+    };
 }