+++ CHANGELOG.md
@@ -9,6 +9,8 @@
  - New awards (Trailblazer, Combat Medic, Dominator, Protector, Finisher, Pain Bringer, Siege Master, Creep Lover, Crowd Controller, Rich Bitch, Al-Qaeda, Pressing Intensifies)
  - DATADRAGON_VERSION=latest & LOCALE env vars
  - Match list minimap
+ - Ability to set the primary user
+ - Persist preferences to localStorage
 
 ## [0.1.0] - 2022-01-15
 ### Initial release